Moved logic of displaying question to own hub
