+++ Arcadia.Ask/Hubs/IQuestionsClient.cs
@@ -11,9 +11,5 @@
         Task QuestionIsRemoved(Guid questionId);
 
         Task QuestionIsVoted(Guid questionId);
-
-        Task DisplayedQuestionChanged(Guid questionId);
-
-        Task DisplayedQuestionHidden();
     }
 }
